Renamed coverage scripts
