+++ .travis.yml
@@ -15,8 +15,8 @@ env:
   # https://docs.travis-ci.com/user/customizing-the-build/#Build-Matrix
   matrix:
     - TEST_COMMAND=build
-    - TEST_COMMAND=coverage-mac
-    - TEST_COMMAND=coverage-windows
+    - TEST_COMMAND=coverage
+    - TEST_COMMAND=coverage-ie
     - TEST_COMMAND=visual
 
 # https://docs.travis-ci.com/user/languages/javascript-with-nodejs#Node.js-v4-(or-io.js-v3)-compiler-requirements
@@ -36,8 +36,8 @@ before_install:
   - npm install -g @blackbaud/skyux-cli
   - export SAVAGE_BRANCH="^savage-"
   - chmod +x ./scripts/travis-script-build.sh
-  - chmod +x ./scripts/travis-script-coverage-mac.sh
-  - chmod +x ./scripts/travis-script-coverage-windows.sh
+  - chmod +x ./scripts/travis-script-coverage.sh
+  - chmod +x ./scripts/travis-script-coverage-ie.sh
   - chmod +x ./scripts/travis-script-deploy.sh
   - chmod +x ./scripts/travis-script-visual.sh
   - chmod +x ./scripts/travis-after-failure.sh
