+++ package.json
@@ -24,8 +24,8 @@
     "test": "npm run test:unit && npm run test:visual",
     "test:ci": "npm run test:unit:ci && npm run test:visual:ci",
     "test:unit": "npm run test:unit:base -- config/karma/local.karma.conf.js",
-    "test:unit:ci:mac": "npm run test:unit:base -- config/karma/ci-mac.karma.conf.js",
-    "test:unit:ci:windows": "npm run test:unit:base -- config/karma/ci-windows.karma.conf.js",
+    "test:unit:ci": "npm run test:unit:base -- config/karma/ci.karma.conf.js",
+    "test:unit:ci:ie": "npm run test:unit:base -- config/karma/ci-ie.karma.conf.js",
     "test:unit:base": "node --max-old-space-size=4096 node_modules/karma/bin/karma start",
     "test:visual": "cd skyux-spa-visual-tests && npm run build-visual && npm run test-visual:local",
     "test:visual:ci": "cd skyux-spa-visual-tests && npm run build-visual && npm run test-visual:ci",
