+++ scripts/travis-script-coverage.sh
@@ -3,7 +3,7 @@ set -e
 
 # Necessary to stop pull requests from forks from running outside of Savage
 if [ "$TRAVIS_SECURE_ENV_VARS" == "true" ]; then
-  npm run test:unit:ci:mac
+  npm run test:unit:ci
   bash <(curl -s https://codecov.io/bash)
 else
   echo -e "Pull requests from forks are run via Savage."
